requirements.txt
matplotlib==3.7.2
numpy==1.23.5
Pillow==9.3.0
Pillow==10.0.0
pygraphviz==1.11
PySimpleGUI==4.60.5
torch==1.9.1+cu111
torchvision==0.10.1+cu111
yacs==0.1.8
